+++ server/wsconn/conn_test.go
@@ -1,4 +1,4 @@
-package server
+package wsconn
 
 import (
 	"net/http"